requirements.txt
aiogram==3.5.0
aiohttp==3.9.5
python-dotenv==1.0.1
